backend/requirements.txt
google-adk>=0.5.0
google-genai>=0.8.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
